Changed the waypoint updater to publish /final_waypoint at a fixed frequency
It was publishing waypoint each time it was getting a /current_pose before, I'm not sure
of the best choice between the two, but this one seems more reasonable at this point.
The frequency is a rosparameter.

Also modified the code to allow a different number of lookahead waypoints (It was a
'parameter' but hardcoded in the code, now it's a ros parameter with the same value)

Minimum effort to prepare for the /traffic_waypoint subscription ;-)